ai_layer/qa_engine.py
import os
from pathlib import Path

import anthropic
from dotenv import load_dotenv

load_dotenv(Path(__file__).parent.parent / ".env")

SYSTEM_PROMPT = """You are an expert on the Hudson Highlands Green Corridors Initiative in Philipstown and Putnam County, NY. Answer questions accurately based on the following context:

THE GREEN CORRIDORS PLAN: Completed in 2022. Led by Hudson Highlands Land Trust (HHLT) with $50,000 funding from NYS DEC Hudson River Estuary Program. Covers towns of Philipstown and Putnam Valley. Identifies priority wildlife habitat connections — forests, marshes, meadows — between existing conserved lands in the eastern Hudson Highlands.

THREE PRIORITY CORRIDOR ZONES IN PHILIPSTOWN:
- Hudson Highlands Core: Western edge along Hudson River, includes state park lands
- Canopus Creek Corridor: Follows Canopus Creek northeast through the town
- Fahnestock Buffer: Eastern zone surrounding Clarence Fahnestock State Park

PRIORITY SPECIES: Bald Eagle (NYS Endangered), Spotted Salamander (Species of Concern), American Eel (Species of Concern), Wood Turtle (NYS Threatened), Bobcat (Species of Concern), Cerulean Warbler (NYS Special Concern).

KEY PARTNERS: Hudson Highlands Land Trust (hhlt.org), Putnam Highlands Audubon Society, Scenic Hudson, NYS DEC, NYS Parks Taconic Commission.

CONSERVATION PROGRAMS AVAILABLE TO PHILIPSTOWN LANDOWNERS: NY State Conservation Easement Tax Credit, USDA Agricultural Conservation Easement Program (ACEP), NYS Environmental Protection Fund (EPF), HHLT conservation easement program.

CONTACT: HHLT Executive Director Katrina Shindledecker, hhlt.org. For municipal planning questions: Philipstown Town Hall 845-265-3329.

Always end answers with: "Source: Hudson Highlands Green Corridors Plan (2022), HHLT."
"""


def answer_question(question: str) -> str:
    api_key = os.environ.get("ANTHROPIC_API_KEY", "")
    if not api_key:
        return "[ANTHROPIC_API_KEY not set — AI Q&A unavailable]"

    client = anthropic.Anthropic(api_key=api_key)
    message = client.messages.create(
        model="claude-sonnet-4-6",
        max_tokens=400,
        system=SYSTEM_PROMPT,
        messages=[{"role": "user", "content": question}],
    )
    return message.content[0].text
