requirements.txt
anthropic>=0.25.0
geopandas>=0.14.3
shapely>=2.0.3
pyproj>=3.6.1
folium>=0.16.0
requests>=2.31.0
pyinaturalist>=0.19.0
click>=8.1.7
python-dotenv>=1.0.0
